embedding/requirements.txt
pandas
numpy
Flask
scikit-learn
pyperclip
python-dotenv
xgboost
openai==0.28.0
